handling backend methods; clear warnings; change router.push

+++ src/main/java/com/nikpetrovic/kanban/model/Card.java
@@ -13,70 +13,70 @@ import java.util.List;
 public class Card {
     private String color;
     private String description;
-    private int id;
+    private long id;
     private String status;
     private List<Task> tasks;
     private String title;
 
     public Card() {
-	super();
+        super();
     }
 
-    public Card(int id, String title, String description, String color, String status, List<Task> tasks) {
-	super();
-	this.id = id;
-	this.title = title;
-	this.description = description;
-	this.color = color;
-	this.status = status;
-	this.tasks = tasks;
+    public Card(long id, String title, String description, String color, String status, List<Task> tasks) {
+        super();
+        this.id = id;
+        this.title = title;
+        this.description = description;
+        this.color = color;
+        this.status = status;
+        this.tasks = tasks;
     }
 
     public String getColor() {
-	return color;
+        return color;
     }
 
     public String getDescription() {
-	return description;
+        return description;
     }
 
-    public int getId() {
-	return id;
+    public long getId() {
+        return id;
     }
 
     public String getStatus() {
-	return status;
+        return status;
     }
 
     public List<Task> getTasks() {
-	return tasks;
+        return tasks;
     }
 
     public String getTitle() {
-	return title;
+        return title;
     }
 
     public void setColor(String color) {
-	this.color = color;
+        this.color = color;
     }
 
     public void setDescription(String description) {
-	this.description = description;
+        this.description = description;
     }
 
-    public void setId(int id) {
-	this.id = id;
+    public void setId(long id) {
+        this.id = id;
     }
 
     public void setStatus(String status) {
-	this.status = status;
+        this.status = status;
     }
 
     public void setTasks(List<Task> tasks) {
-	this.tasks = tasks;
+        this.tasks = tasks;
     }
 
     public void setTitle(String title) {
-	this.title = title;
+        this.title = title;
     }
 }
